New Models added for form no. 3.11 to 3.13 including all other detail models

+++ WrpCcNocWeb/Models/CcModule/CcModAppProject_31_IndvDetail.cs
@@ -91,11 +91,11 @@ namespace WrpCcNocWeb.Models
 
         [Column("CropProduction", Order = 17)]
         [Display(Name = "Crop Production")]
-        public int? CropProduction { get; set; }
+        public double? CropProduction { get; set; }
 
         [Column("FishProduction", Order = 18)]
         [Display(Name = "Fish Production")]
-        public int? FishProduction { get; set; }
+        public double? FishProduction { get; set; }
 
         [Column("FishDiversity", Order = 19)]
         [Display(Name = "Fish Diversity")]
