+++ WrpCcNocWeb/Models/CcModule/CcModAppProject_33_IndvDetail.cs
@@ -128,11 +128,11 @@ namespace WrpCcNocWeb.Models
 		
 		[Column("CropProduction", Order = 26)]
         [Display(Name = "Crop Production")]
-        public int? CropProduction { get; set; }
+        public double? CropProduction { get; set; }
 		
         [Column("FishProduction", Order = 27)]
         [Display(Name = "Fish Production")]		
-        public int? FishProduction { get; set; }
+        public double? FishProduction { get; set; }
 
         [Column("FishDiversity", Order = 28)]
         [Display(Name = "Fish Diversity")]
